Shortcuts editor layout
